WIP: Adds gtest and refactors DrawLocation
We're in the process of adding GTest to the codebase. Struggling with
understanding targets and how to ensure that the soh target does not get
tests compiled into it (tests might affect global state).

This is because we started undertaking a refactor in DrawLocation that
changes how the text for a check is rendered; simplifying if statements
and breaking pieces of functionality out into smaller functions. But we
didn't feel confident in our refactor, so we wanted to add tests to the
project.

Co-authored-by: Michael Stergianis <michael.stergianis@gmail.com>